+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.2.0] - 2019-02-12
+
+### Added
+
+- Encode COU role title and affiliation as role information in generated entitlements
+- Encode COU admins group member and owner role information in generated entitlements
+
 ## [v1.1.1] - 2019-01-22
 
 ### Fixed
@@ -30,6 +37,7 @@ This version is compatible with [SimpleSAMLphp v1.14](https://simplesamlphp.org/
   - urnNamespace
   - urnAuthority
   - COmanage registry redirect URLs
+  - voRoles
 
 ### Fixed
 - sql query failure due to typo error